tempory finished websocket mini Project
mini project with websocket server and client side in mind.

allows you to message in real time to a server and retrieve the message
back to the client. this is beneficial for apps such as instant
messaging. may turn it into exactly that in the future
following tutorials online such as
https://www.w3schools.com/nodejs/nodejs_websockets.asp